+++ api/src/universal_inbox/task/service.rs
@@ -6,14 +6,15 @@ use std::{
 
 use anyhow::{anyhow, Context};
 use chrono::{TimeDelta, Utc};
+use slack_morphism::prelude::{SlackEventCallbackBody, SlackPushEventCallback};
 use sqlx::{Postgres, Transaction};
 use tokio::sync::RwLock;
 use tracing::{debug, error, info};
 
 use universal_inbox::{
     notification::{
-        service::NotificationPatch, Notification, NotificationSource, NotificationSourceKind,
-        NotificationStatus,
+        service::NotificationPatch, Notification, NotificationMetadata, NotificationSource,
+        NotificationSourceKind, NotificationStatus,
     },
     task::{
         service::TaskPatch, ProjectSummary, Task, TaskCreation, TaskCreationResult, TaskId,
@@ -24,12 +25,15 @@ use universal_inbox::{
 };
 
 use crate::{
-    integrations::{task::TaskSourceService, todoist::TodoistService},
+    integrations::{
+        notification::NotificationSourceService, slack::SlackService, task::TaskSourceService,
+        todoist::TodoistService,
+    },
     repository::{task::TaskRepository, Repository},
     universal_inbox::{
         integration_connection::service::IntegrationConnectionService,
         notification::service::NotificationService, user::service::UserService,
-        UniversalInboxError, UpdateStatus,
+        UniversalInboxError, UpdateStatus, UpsertStatus,
     },
 };
 
@@ -37,6 +41,7 @@ pub struct TaskService {
     repository: Arc<Repository>,
     todoist_service: TodoistService,
     notification_service: Weak<RwLock<NotificationService>>,
+    slack_service: SlackService,
     integration_connection_service: Arc<RwLock<IntegrationConnectionService>>,
     user_service: Arc<RwLock<UserService>>,
     min_sync_tasks_interval_in_minutes: i64,
@@ -47,6 +52,7 @@ impl TaskService {
         repository: Arc<Repository>,
         todoist_service: TodoistService,
         notification_service: Weak<RwLock<NotificationService>>,
+        slack_service: SlackService,
         integration_connection_service: Arc<RwLock<IntegrationConnectionService>>,
         user_service: Arc<RwLock<UserService>>,
         min_sync_tasks_interval_in_minutes: i64,
@@ -55,6 +61,7 @@ impl TaskService {
             repository,
             todoist_service,
             notification_service,
+            slack_service,
             integration_connection_service,
             user_service,
             min_sync_tasks_interval_in_minutes,
@@ -97,6 +104,53 @@ impl TaskService {
             .update_task(executor, &task.source_id, patch, user_id)
             .await?;
 
+        Ok(())
+    }
+
+    #[tracing::instrument(
+        level = "debug",
+        skip(self, executor, task, notifications),
+        fields(
+            task_id = task.id.to_string(),
+            notification_ids = notifications.iter().map(|n| n.id.to_string()).collect::<Vec<String>>().join(", ")
+        ),
+        err
+    )]
+    pub async fn apply_synced_task_side_effect<'a>(
+        &self,
+        executor: &mut Transaction<'a, Postgres>,
+        task: &Task,
+        notifications: Vec<Notification>,
+        user_id: UserId,
+    ) -> Result<(), UniversalInboxError> {
+        for notification in notifications {
+            debug!(
+                "Applying side effect for task {} and notification {}",
+                task.id, notification.id
+            );
+            match &notification.metadata {
+                NotificationMetadata::Slack(box SlackPushEventCallback {
+                    event: SlackEventCallbackBody::StarAdded(_),
+                    ..
+                })
+                | NotificationMetadata::Slack(box SlackPushEventCallback {
+                    event: SlackEventCallbackBody::StarRemoved(_),
+                    ..
+                }) => {
+                    if task.status == TaskStatus::Done {
+                        self.slack_service
+                            .delete_notification_from_source(executor, &notification, user_id)
+                            .await?;
+                    } else if task.status == TaskStatus::Active {
+                        self.slack_service
+                            .undelete_notification_from_source(executor, &notification, user_id)
+                            .await?;
+                    }
+                }
+
+                _ => {}
+            }
+        }
         Ok(())
     }
 
@@ -183,7 +237,7 @@ impl TaskService {
 
         Ok(Box::new(TaskCreationResult {
             task: *task,
-            notification: notification.map(|n| *n),
+            notifications: notification.into_iter().map(|n| *n).collect(),
         }))
     }
 
@@ -219,6 +273,7 @@ impl TaskService {
         Ok(created_task)
     }
 
+    #[tracing::instrument(level = "debug", skip(self, executor, task_source_service), err)]
     async fn sync_source_tasks_and_notifications<'a, T: Debug, U>(
         &self,
         executor: &mut Transaction<'a, Postgres>,
@@ -273,7 +328,7 @@ impl TaskService {
             .await?;
         match task_source_service.fetch_all_tasks(executor, user_id).await {
             Ok(source_tasks) => {
-                let tasks = self
+                let upsert_tasks = self
                     .save_tasks_from_source(
                         executor,
                         task_source_service,
@@ -283,18 +338,71 @@ impl TaskService {
                     )
                     .await?;
 
-                let mut notifications_by_task_id: HashMap<Option<TaskId>, Notification> =
-                    if integration_connection
-                        .provider
-                        .should_create_notification_from_inbox_task()
-                    {
-                        // Create/update notifications for tasks in the Inbox
-                        let tasks_in_inbox: Vec<Task> = tasks
-                            .iter()
-                            .filter(|task| task.is_in_inbox())
-                            .cloned()
-                            .collect();
+                let created_or_updated_tasks: Vec<Task> = upsert_tasks
+                    .iter()
+                    .filter_map(|upsert_task| match upsert_task {
+                        UpsertStatus::Created(task) | UpsertStatus::Updated(task) => {
+                            Some((**task).clone())
+                        }
+                        _ => None,
+                    })
+                    .collect();
 
+                let mut notifications_by_task_id: HashMap<TaskId, Vec<Notification>> =
+                    HashMap::new();
+                for task in &created_or_updated_tasks {
+                    let notifications = self
+                        .notification_service
+                        .upgrade()
+                        .context("Unable to access notification_service from task_service")?
+                        .read()
+                        .await
+                        .list_notifications(executor, vec![], true, Some(task.id), None, user_id)
+                        .await?;
+
+                    notifications_by_task_id.insert(
+                        task.id,
+                        notifications
+                            .content
+                            .into_iter()
+                            .map(|notification_with_task| notification_with_task.into())
+                            .collect(),
+                    );
+                }
+
+                let (tasks_in_inbox, tasks_not_in_inbox): (Vec<Task>, Vec<Task>) =
+                    created_or_updated_tasks
+                        .into_iter()
+                        .partition(|task| task.is_in_inbox());
+
+                if integration_connection
+                    .provider
+                    .should_create_notification_from_inbox_task()
+                {
+                    let notification_source_kind =
+                        task_source_service.get_notification_source_kind();
+                    // Create notifications from tasks in the inbox if there is no existing notification
+                    // for this task or if there is an existing notification with the same source kind
+                    let notifications_from_tasks: Vec<Notification> = tasks_in_inbox
+                        .into_iter()
+                        .filter_map(|task| {
+                            if let Some(existing_notifications) =
+                                notifications_by_task_id.get(&task.id)
+                            {
+                                if !existing_notifications.is_empty()
+                                    && existing_notifications
+                                        .iter()
+                                        .all(|n| n.get_source_kind() != notification_source_kind)
+                                {
+                                    return None;
+                                }
+                            }
+                            Some(task.into_notification(user_id))
+                        })
+                        .collect();
+
+                    if !notifications_from_tasks.is_empty() {
+                        // Create/update notifications for tasks in the Inbox
                         let upsert_inbox_notifications = self
                             .notification_service
                             .upgrade()
@@ -303,31 +411,36 @@ impl TaskService {
                             .await
                             .save_notifications_from_source(
                                 executor,
-                                task_source_service.get_notification_source_kind(),
-                                tasks_in_inbox
-                                    .into_iter()
-                                    .map(|task| task.into_notification(user_id))
-                                    .collect(),
+                                notification_source_kind,
+                                notifications_from_tasks,
                                 true,
                                 task_source_service.is_supporting_snoozed_notifications(),
                                 user_id,
                             )
                             .await?;
 
-                        upsert_inbox_notifications
-                            .into_iter()
-                            .map(|upsert_notification| {
-                                let notification = upsert_notification.value();
-                                (notification.task_id, *notification)
-                            })
-                            .collect()
-                    } else {
-                        HashMap::new()
-                    };
+                        for upsert_notification in upsert_inbox_notifications.into_iter() {
+                            let notification = upsert_notification.value();
+                            if let Some(task_id) = notification.task_id {
+                                let notifications_for_task =
+                                    notifications_by_task_id.entry(task_id).or_default();
+                                if let Some(index) = notifications_for_task
+                                    .iter()
+                                    .position(|n| n.id == notification.id)
+                                {
+                                    notifications_for_task[index] = *notification;
+                                } else {
+                                    notifications_for_task.push(*notification);
+                                }
+                            }
+                        }
+                    }
+                }
 
                 // Update existing notifications for tasks that are not in the Inbox anymore
-                for task in tasks.iter().filter(|task| !task.is_in_inbox()) {
-                    self.notification_service
+                for task in tasks_not_in_inbox {
+                    let updated_notifications = self
+                        .notification_service
                         .upgrade()
                         .context("Unable to access notification_service from task_service")?
                         .read()
@@ -342,18 +455,44 @@ impl TaskService {
                             },
                         )
                         .await?;
+                    for updated_notification in updated_notifications.into_iter() {
+                        let notifications_for_task =
+                            notifications_by_task_id.entry(task.id).or_default();
+                        if updated_notification.updated {
+                            if let Some(notification) = updated_notification.result {
+                                if let Some(index) = notifications_for_task
+                                    .iter()
+                                    .position(|n| n.id == notification.id)
+                                {
+                                    notifications_for_task[index] = notification;
+                                }
+                            }
+                        }
+                    }
                 }
 
-                Ok(tasks
-                    .into_iter()
-                    .map(move |task| {
-                        let task_id = task.id;
-                        TaskCreationResult {
-                            task,
-                            notification: notifications_by_task_id.remove(&Some(task_id)),
-                        }
-                    })
-                    .collect())
+                let mut tasks_creation_result = vec![];
+                for upsert_task in upsert_tasks {
+                    let task = upsert_task.value();
+                    let notifications = notifications_by_task_id.remove(&task.id);
+
+                    if let Some(notifications) = &notifications {
+                        self.apply_synced_task_side_effect(
+                            executor,
+                            &task,
+                            notifications.to_vec(),
+                            user_id,
+                        )
+                        .await?;
+                    }
+
+                    tasks_creation_result.push(TaskCreationResult {
+                        task: *task,
+                        notifications: notifications.unwrap_or_default(),
+                    });
+                }
+
+                Ok(tasks_creation_result)
             }
             Err(e) => {
                 self.integration_connection_service
@@ -395,25 +534,29 @@ impl TaskService {
         source_tasks: &[T],
         is_incremental_update: bool,
         user_id: UserId,
-    ) -> Result<Vec<Task>, UniversalInboxError> {
-        let mut tasks = vec![];
+    ) -> Result<Vec<UpsertStatus<Box<Task>>>, UniversalInboxError> {
+        let mut upsert_tasks = vec![];
         for source_task in source_tasks {
             let task = task_source_service
                 .build_task(executor, source_task, user_id)
                 .await?;
-            let uptodate_task = self
+            let upsert_task = self
                 .repository
                 .create_or_update_task(executor, task)
                 .await?;
-            tasks.push(uptodate_task);
+            upsert_tasks.push(upsert_task);
         }
+        info!(
+            "{} Todoist tasks successfully synced for user {user_id}.",
+            upsert_tasks.len()
+        );
 
         // For incremental synchronization tasks services, there is no need to update stale tasks
         // Not yet used as Todoist is incremental
         if !is_incremental_update {
-            let source_task_ids = tasks
+            let source_task_ids = upsert_tasks
                 .iter()
-                .map(|task| task.source_id.clone())
+                .map(|upsert_task| upsert_task.value_ref().source_id.clone())
                 .collect::<Vec<String>>();
 
             self.repository
@@ -427,7 +570,7 @@ impl TaskService {
                 .await?;
         }
 
-        Ok(tasks)
+        Ok(upsert_tasks)
     }
 
     #[tracing::instrument(level = "debug", skip(self, executor), err)]