+++ api/tests/api/main.rs
@@ -8,6 +8,7 @@ mod test_linear_notifications;
 mod test_misc;
 mod test_notifications;
 mod test_slack_notifications;
+mod test_slack_tasks;
 mod test_slack_webhook;
 mod test_sync_github_notifications;
 mod test_sync_google_mail_threads;